Fix set_staff.
